backend/rest_server/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
boto3==1.40.55
botocore==1.40.55
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==46.0.2
distro==1.9.0
Django==5.2.7
django-environ==0.12.0
django-oauth-toolkit==3.1.0
django-storages==1.14.6
djangorestframework==3.16.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.11.1
jmespath==1.0.1
jwcrypto==1.5.6
oauthlib==3.3.1
openai==2.5.0
pillow==12.0.0
psycopg==3.2.10
psycopg-binary==3.2.10
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
PyMySQL==1.1.2
python-dateutil==2.9.0.post0
requests==2.32.5
s3transfer==0.14.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
